+++ src/js/sagas/deleteExperienceSaga.js
@@ -5,12 +5,15 @@ import { push } from "react-router-redux";
 
 export default function* deleteExperienceSaga(action) {
   const id = action.payload;
-  yield put(actions.deleteExperienceAttempt());
-  try {
-    yield call(api.deleteExperienceApi, id);
-    yield put(actions.deleteExperienceSuccess());
-    yield put(push("/dashboard"));
-  } catch (e) {
-    yield put(actions.deleteExperienceError(e.response.data));
+  if (confirm("Are you confirm to permanent delete Experience")) {
+    yield put(actions.deleteExperienceAttempt());
+    try {
+      yield call(api.deleteExperienceApi, id);
+      yield put(actions.deleteExperienceSuccess());
+      alert("Experience Deleted Successfuly");
+      yield put(push("/dashboard"));
+    } catch (e) {
+      yield put(actions.deleteExperienceError(e.response.data));
+    }
   }
 }
