+++ src/js/sagas/deleteProfileDataSaga.js
@@ -4,12 +4,15 @@ import * as api from "../api";
 import { push } from "react-router-redux";
 
 export default function* deleteProfileDataSaga() {
-  yield put(actions.deleteProfileDataAttempt());
-  try {
-    yield call(api.deleteProfileApi);
-    yield put(actions.deleteProfileDataSuccess());
-    yield put(push("/"));
-  } catch (e) {
-    yield put(actions.deleteProfileDataError(e.response.data));
+  if (confirm("Are you sure to permanent delete Profile")) {
+    yield put(actions.deleteProfileDataAttempt());
+    try {
+      yield call(api.deleteProfileApi);
+      yield put(actions.deleteProfileDataSuccess());
+      alert("Profile Deleted Successfuly");
+      yield put(push("/"));
+    } catch (e) {
+      yield put(actions.deleteProfileDataError(e.response.data));
+    }
   }
 }
